# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: testimonials
    description: Testimonial operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: size
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonials]
      summary: List testimonials
      responses:
        "200":
          description: Testimonials list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
  /products/summary:
    get:
      operationId: getProductsSummary
      tags: [products]
      summary: Get products summary with counts by category
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsSummary"
  /products/new-arrivals:
    get:
      operationId: listNewArrivals
      tags: [products]
      summary: Get newest products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 4
      responses:
        "200":
          description: New arrivals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/on-sale:
    get:
      operationId: listOnSaleProducts
      tags: [products]
      summary: Get products on sale
      responses:
        "200":
          description: Products on sale
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        salePrice:
          type: number
          nullable: true
        categoryId:
          type: integer
        categoryName:
          type: string
        images:
          type: array
          items:
            type: string
        sizes:
          type: array
          items:
            type: string
        featured:
          type: boolean
        isNew:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - price
        - categoryId
        - categoryName
        - images
        - sizes
        - featured
        - isNew
        - createdAt
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        image:
          type: string
        productCount:
          type: integer
      required:
        - id
        - name
        - slug
        - image
        - productCount
    Testimonial:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        comment:
          type: string
        rating:
          type: integer
        avatar:
          type: string
      required:
        - id
        - name
        - comment
        - rating
        - avatar
    ProductsSummary:
      type: object
      properties:
        totalProducts:
          type: integer
        totalCategories:
          type: integer
        onSaleCount:
          type: integer
        categoryCounts:
          type: array
          items:
            type: object
            properties:
              categoryName:
                type: string
              count:
                type: integer
            required:
              - categoryName
              - count
      required:
        - totalProducts
        - totalCategories
        - onSaleCount
        - categoryCounts