# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  salePrice?: number | null;
  categoryId: number;
  categoryName: string;
  images: string[];
  sizes: string[];
  featured: boolean;
  isNew: boolean;
  createdAt: Date;
}