# lib/api-zod/src/generated/types/productsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductsSummaryCategoryCountsItem } from "./productsSummaryCategoryCountsItem";

export interface ProductsSummary {
  totalProducts: number;
  totalCategories: number;
  onSaleCount: number;
  categoryCounts: ProductsSummaryCategoryCountsItem[];
}